Added SNB model using code from master branch
Option to use SNB is controlled by the snb_model boolean in BOUT.inp, creating a HeatFluxSNB object. This changes ddt(electrons.P) in the heat_conduction section of the electron model. Note, there is still a factor of 2 in ddt(electrons.P) which I think should be removed in future.