SD1D: Removed factor of 2 in heat conduction update.
Currently this model assumes instant equipartition of energy between
electrons and ions. The factor of 2 was previously used to account for
this, with a factor of 0.5 in the ddt(ions.P) update. I've moved this factor of 0.5
to the ddt(electrons.P) update, meaning of the energy change due to heat
conduction, half the change goes to the electrons, and an equal amount goes
to the ions. H