Swapped round ion pressure and momentum updates so momentum is updated even without heat conduction.
First, I realised I got the last update slightly wrong - the contribution to ion momentum from ion pressure is already included in species.cxx so I only need to add the contribution from electron pressure.
Next, I want this contribution to be included even if heat conduction is turned off, so I moved this to outside the if(heat_conduction) clause.
Finally, I only need to update the ion pressure if there is heat conduction, so I moved this inside the if(heat_conduction) clause